LayoutTests/js/dom/lockdown-mode/XSLTProcessor-disabled.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Ensures XSLTProcessor is unavailable to JavaScript in LDM.");

shouldBeTrue("typeof(XSLTProcessor) === 'undefined'");
</script>
</body>
</html>
