LayoutTests/js/dom/lockdown-mode/XSLTProcessor-disabled-expected.txt

Ensures XSLTProcessor is unavailable to JavaScript in LDM.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof(XSLTProcessor) === 'undefined' is true
PASS successfullyParsed is true

TEST COMPLETE
